Adding higher order component to control header/drawer
